mbg-api/app/requirements.txt
fastapi
fastapi[standard]
psycopg2-binary
pydantic-settings
requests
sqlalchemy
